fix dashboard access

+++ ecommerce-frontend/src/pages/admin/Customers.tsx
@@ -53,13 +53,14 @@ const column: Column<DataType>[] = [
 const Search = () => {
   const { user } = useSelector((state: RootState) => state.userReducer);
 
-  const { isLoading, data, isError, error } = useAllUsersQuery(user!._id!);
+  const { isLoading, data, isError, error } = useAllUsersQuery("");
 
   const [rows, setRows] = useState<DataType[]>([]);
 
   const [deleteUser] = useDeleteUserMutation();
 
   const deleteHandler = async (userId: string) => {
+    if(user?.role!="admin") return toast.error("You're Not admin");
     const res = await deleteUser({ userId, adminUserId: user!._id! });
     responseToast(res, null, "");
   };
